requirements.txt
astroid==1.4.6
beautifulsoup4==4.4.1
bs4==0.0.1
colorama==0.3.7
future==0.15.2
lazy-object-proxy==1.2.2
pylint==1.5.6
python-telegram-bot==4.2.0
requests==2.10.0
six==1.10.0
wrapt==1.10.8
